% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcrdataportal-package.R
\docType{package}
\name{unhcrdataportal-package}
\alias{unhcrdataportal}
\alias{unhcrdataportal-package}
\title{unhcrdataportal: Retrieve Data From UNHCR Data Portal}
\description{
Use a function to retrieve the catalog / data in the system.
}
\author{
\strong{Maintainer}: Hisham Galal \email{galalh@unhcr.org}

Authors:
\itemize{
  \item Edouard Legoupil \email{legoupil@unhcr.org}
}

}
\keyword{internal}
